website/requirements.txt
Django==1.11.9
django-classy-tags==0.8.0
django-cms==3.4.5
django-filer==1.3.0
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==1.3
django-sekizai==0.10.0
django-treebeard==4.2.0
django-parler==1.9.1
django-object-actions==0.10.0
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-audio==1.0.2
djangocms-file==2.0.2
djangocms-googlemap==1.1.1
djangocms-icon==0.2.0
djangocms-link==2.1.2
djangocms-picture==2.0.6
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
easy-thumbnails==2.4.2
html5lib==0.9999999
Pillow==5.0.0
pytz==2017.3
six==1.11.0
Unidecode==0.4.21
aldryn-apphooks-config==0.3.3
